requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
Pillow
numpy
opencv-python
face-recognition
dlib
cmake